feature/point_reason 테이블 조작 기능 완료 #1. insert, delete, update 테스트 완료 #2. 참조키 무결성 충족

+++ src/main/java/rev/team/POINT_SERVICE/service/PointReasonService.java
@@ -47,13 +47,17 @@ public class PointReasonService {
     }
 
     public String delete(Long id) {
-        pointHistoryRepository.deleteById(id);
+        if (!pointHistoryRepository.findById(id).isEmpty()) {
+            pointHistoryRepository.deleteById(id);
+        }
         pointReasonRepository.deleteById(id);
         return "DELETE SUCCESS";
     }
 
     public String update(Long id, PointDTO pointDTO) {
-        pointHistoryRepository.updateById(id, pointDTO.getPointReason(), pointDTO.getPoint());
+        if (!pointHistoryRepository.findById(id).isEmpty()) {
+            pointHistoryRepository.updateById(id, pointDTO.getPointReason(), pointDTO.getPoint());
+        }
         pointReasonRepository.updateById(id, pointDTO.getPointReason(), pointDTO.getPoint());
         return "UPDATE SUCCESS";
     }
